U. S.     O P E N
This was a rollercoaster of a day.

We started out thinking that IFT (a really good Chinese team) was going
to pick us, but then we screwed up a match in front of them. Our autons
never worked in quals, so we ended up ranked 5th for alliance selection
planning to pick X team.

Then we ran skills. If our mobile goal intake had stayed out or we had
driven 5 inches less on our last mobile goal (like our code said to do
yesterday) we would have won, but as it is we got second.

IFT was ranked first and picked 6135W, making a driver-loader/fielder
combination that I didn't think could be beaten. X team was picked by
the 2nd seed team. Because there were so few capable teams, lots of the
highly ranked bots in the competition were terrible and the two ahead of
us elimininated most of our other possible choices by getting declined
about a dozen times. I thought we screwed ourselves over by picking
3815M over 1008M in the first round, but 3815 turned out to be really
versatile and had a much more reliable goal intake. We were completely
unprepared for the second round of picking because Ted and Ben had been
scouting under the assumption that there would be 8 alliances so we
ended up with the complete toaster 2501B.

And then we made it to division finals and beat IFT and 6135 without a
third match.

And then we lost by one point in finals 3 after 3815's drive burned out
and they didn't park or score the last mobile goal.